+++ CHANGELOG.md
@@ -57,11 +57,12 @@
 * [CHANGE] Compactor: removed overlapping sources detection. Overlapping sources may exist due to edge cases (timing issues) when horizontally sharding compactor with `split-and-merge` strategy, but are correctly handled by compactor. #494
 * [CHANGE] Rename metric `cortex_query_fetched_chunks_bytes_total` to `cortex_query_fetched_chunk_bytes_total` to be consistent with the limit name. #476
 * [CHANGE] The `status_code` label on gRPC client metrics has changed from '200' and '500' to '2xx', '5xx', '4xx', 'cancel' or 'error'. #537
-* [CHANGE] Remove chunks storage engine. #510 #545 #743 #744 #748 #753 #755 #757 #758 #759
+* [CHANGE] Remove chunks storage engine. #510 #545 #743 #744 #748 #753 #755 #757 #758 #759 #762
   * The following CLI flags (and their respective YAML config options) have been removed:
     * `-store.engine`
     * `-ingester.checkpoint-duration`
     * `-ingester.checkpoint-enabled`
+    * `-ingester.chunk-encoding`
     * `-ingester.chunk-age-jitter`
     * `-ingester.concurrent-flushes`
     * `-ingester.flush-on-shutdown-with-wal-enabled`
@@ -100,6 +101,10 @@
     * All `-grpc-store.*` flags
   * The following API endpoints have been removed:
     * `/api/v1/chunks` and `/chunks`
+  * The following metrics have been removed:
+    * `prometheus_local_storage_chunk_ops_total`
+    * `prometheus_local_storage_chunkdesc_ops_total`
+    * `prometheus_local_storage_memory_chunkdescs`
 * [CHANGE] Compactor: compactor now uses deletion marks from `<tenant>/markers` location in the bucket. Marker files are no longer fetched, only listed. #550
 * [CHANGE] Compactor: Default value of `-compactor.block-sync-concurrency` has changed from 20 to 8. This flag is now only used to control number of goroutines for downloading and uploading blocks during compaction. #552
 * [CHANGE] Memberlist: changed probe interval from `1s` to `5s` and probe timeout from `500ms` to `2s`. #563